+++ bin/fuel-core/src/cli/run.rs
@@ -127,7 +127,7 @@ pub struct Command {
     /// The default value is 7 days = 604800 blocks.
     ///
     /// The `BlockHeight` is `u32`, meaning the maximum possible number of blocks
-    /// is less than 137 years. `2^32 / 24 / 60 / 60/ 365` = `136.1925195332` years.
+    /// is less than 137 years. `2^32 / 24 / 60 / 60 / 365` = `136.1925195332` years.
     /// If the value is 136 years or more, the rewind feature is enabled for all blocks.
     #[clap(long = "state-rewind-duration", default_value = "7d", env)]
     pub state_rewind_duration: humantime::Duration,
@@ -140,6 +140,10 @@ pub struct Command {
     /// configuration.
     #[arg(name = "DB_PRUNE", long = "db-prune", env, default_value = "false")]
     pub db_prune: bool,
+
+    /// The determines whether to continue the services on internal error or not.
+    #[clap(long = "continue-services-on-error", default_value = "false", env)]
+    pub continue_on_error: bool,
 
     /// Should be used for local development only. Enabling debug mode:
     /// - Allows GraphQL Endpoints to arbitrarily advance blocks.
@@ -236,6 +240,7 @@ impl Command {
             state_rewind_duration,
             db_prune,
             snapshot,
+            continue_on_error,
             vm_backtrace,
             debug,
             utxo_validation,
@@ -384,6 +389,7 @@ impl Command {
             snapshot_reader,
             debug,
             native_executor_version,
+            continue_on_error,
             utxo_validation,
             block_production: trigger,
             vm: VMConfig {