+++ crates/fuel-core/src/service/adapters.rs
@@ -30,7 +30,9 @@ pub mod block_importer;
 pub mod consensus_module;
 pub mod consensus_parameters_provider;
 pub mod executor;
+pub mod fuel_gas_price_provider;
 pub mod graphql_api;
+pub mod import_result_provider;
 #[cfg(feature = "p2p")]
 pub mod p2p;
 pub mod producer;
@@ -39,8 +41,6 @@ pub mod relayer;
 #[cfg(feature = "p2p")]
 pub mod sync;
 pub mod txpool;
-
-pub mod fuel_gas_price_provider;
 
 #[derive(Debug, Clone)]
 pub struct ConsensusParametersProvider {